Changes made that were requested by Nabil: moving the read_suvi function to io.py, documentation in one place only, SOLAR_CLASES variables etc up in the code.
