+++ docs/code_ref/index.rst
@@ -13,4 +13,5 @@ Code Reference
    iris
    lyra
    rhessi
+   io
    suvi
