+++ docs/code_ref/io.rst
@@ -1,0 +1,1 @@
+.. automodapi:: sunkit_instruments.io
