+++ sunkit_instruments/io/__init__.py
@@ -1,0 +1,1 @@
+from .io import *  # NOQA
